Rehabilitate Credit Notes and fix journal entry numbering collision
Credit Notes (FTA-compliant invoice corrections) out of quarantine:
- credit_notes + credit_note_lines tables, insert schemas, types;
  idempotent migration 0060
- Eight Drizzle storage methods plus createJournalEntryWithLines (atomic
  entry-number generation + balanced insert) for document flows that post
  entries as a side effect
- Issue flow posts the reversing journal entry (debit revenue + VAT
  output, credit A/R) and links it via journalEntryId; void flow guarded
- Account lookups now use stable ACCOUNT_CODES with name fallbacks — the
  default UAE chart has no account literally named 'Sales Revenue', so
  the previous name-only lookup 500'd for every company on the seed
- ISO date strings coerced at the route boundary; routes registered;
  /credit-notes routed with a Sales sidebar entry
- E2E extended: route crawl + lifecycle flow (create -> renders -> issue
  -> assert issued status and journal link)

CRITICAL ledger fix found by the new flow: journal entry numbering was
broken for every entry after the first of a day. Bound parameters arrive
as text, and SUBSTRING(entry_number FROM <text>) is the REGEX form of
substring — it returned NULL for every row, so MAX was always 0 and every
generated number was -001; the second entry of the day failed the unique
constraint with a 500. Fixed with ::int casts at all three sites
(generateEntryNumber, the transactional generator, fixed-assets batch
numbering). The journal + credit-note E2E flows now post two entries on
the same day, locking in the regression.

Verified: 69-route crawl + 4 flows green, 491 unit tests, typecheck,
hygiene checks, production build.

+++ server/routes/fixed-assets.routes.ts
@@ -34,7 +34,7 @@ async function makeEntryNumberAllocator(
   const likePattern = prefix + '-%';
 
   const result = await client.query(
-    `SELECT COALESCE(MAX(CAST(SUBSTRING(entry_number FROM $1) AS INTEGER)), 0) AS max_seq
+    `SELECT COALESCE(MAX(CAST(SUBSTRING(entry_number FROM $1::int) AS INTEGER)), 0) AS max_seq
        FROM journal_entries
       WHERE company_id = $2 AND entry_number LIKE $3`,
     [counterStart, companyId, likePattern],
